+++ src/prelude.rs
@@ -5,7 +5,6 @@
 //! ```
 
 pub use crate::error::{Error, Result};
-pub use crate::files::Files;
 #[cfg(feature = "stream")]
 pub use crate::stream::Streaming;
-pub use crate::{Client, ClientBuilder, Endpoint};
+pub use crate::{Client, ClientBuilder, Endpoint, Multipart, Response};